commit changes for CAN
CAN function

+++ SLIDE-x-BENCH/RECIPE/bs/frst.c
@@ -43,7 +43,7 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef double TARGET_TYPE;
+typedef int8_t TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
 //void prototype(int8_t n, int8_t arr[n], int8_t key);
